+++ lib/__tests__/transform.js
@@ -111,31 +111,20 @@ describe('classic components', function() {
     expect(generateSnapshot(source, template)).toMatchSnapshot();
   });
 
-test('handles `ariaRole` correctly', () => {
-  let source = `
+  test('handles `ariaRole` correctly', () => {
+    let source = `
     export default Component.extend({
       ariaRole: 'button',
     });
   `;
 
-  let template = `foo`;
+    let template = `foo`;
 
-  expect(generateSnapshot(source, template)).toMatchSnapshot();
-});
+    expect(generateSnapshot(source, template)).toMatchSnapshot();
+  });
 
-test('throws if `Component.extend({ ... })` is not found', () => {
-  let source = `
-    export default class extends Component {
-    }
-  `;
-
-  expect(() => transform(source, '')).toThrowErrorMatchingInlineSnapshot(
-    `"Unsupported component type. Only classic components (\`Component.extend({ ... }\`) are supported currently."`
-  );
-});
-
-test('throws if `Component.extend({ ... })` argument is not found', () => {
-  let source = `
+  test('throws if `Component.extend({ ... })` argument is not found', () => {
+    let source = `
     export default Component.extend();
   `;
 
@@ -215,8 +204,8 @@ test('throws if `Component.extend({ ... })` argument is not found', () => {
     );
   });
 
-test('throws if component is using a computed property for `ariaRole`', () => {
-  let source = `
+  test('throws if component is using a computed property for `ariaRole`', () => {
+    let source = `
     export default Component.extend({
       ariaRole: computed(function() {
         return 'button';
@@ -224,13 +213,13 @@ test('throws if component is using a computed property for `ariaRole`', () => {
     });
   `;
 
-  expect(() => transform(source, '')).toThrowErrorMatchingInlineSnapshot(
-    `"Codemod does not support computed properties for \`ariaRole\`."`
-  );
-});
+    expect(() => transform(source, '')).toThrowErrorMatchingInlineSnapshot(
+      `"Codemod does not support computed properties for \`ariaRole\`."`
+    );
+  });
 
-test('multi-line template', () => {
-  let source = `export default Component.extend({});`;
+  test('multi-line template', () => {
+    let source = `export default Component.extend({});`;
 
     let template = `
 {{#if this.foo}}
