+++ lib/transform/native.js
@@ -122,16 +122,6 @@ module.exports = function transformNativeComponent(root, options) {
   // set `@tagName('')`
   addClassDecorator(exportDefaultDeclaration, 'tagName', [j.stringLiteral('')]);
   ensureImport(root, 'tagName', '@ember-decorators/component');
-  // let tagNamePath = j(classBody)
-  //   .find(j.ClassProperty)
-  //   // .filter(path => path.parentPath === properties)
-  //   .filter(path => isProperty(path, 'tagName'));
-  //
-  // if (tagNamePath.length === 1) {
-  //   j(tagNamePath.get('value')).replaceWith(j.stringLiteral(''));
-  // } else {
-  //   classBody.unshift(j.classProperty(j.identifier('tagName'), j.stringLiteral('')));
-  // }
 
   // remove `elementId`, `attributeBindings`, `classNames` and `classNameBindings`
   j(classBody)
