+++ lib/transform/template.js
@@ -1,4 +1,4 @@
-const { indentLines } = require('../utils/classic');
+const { indentLines } = require('../utils/template');
 
 const templateRecast = require('ember-template-recast');
