+++ lib/utils/classic.js
@@ -91,13 +91,6 @@ function findClassNameBindings(properties) {
   return classNameBindings;
 }
 
-function indentLines(content) {
-  return content
-    .split('\n')
-    .map(it => `  ${it}`)
-    .join('\n');
-}
-
 module.exports = {
   isProperty,
   isMethod,
@@ -109,5 +102,4 @@ module.exports = {
   findClassNameBindings,
   findElementId,
   findTagName,
-  indentLines,
 };
