+++ lib/utils/native.js
@@ -178,13 +178,6 @@ function removeDecorator(root, classDeclaration, name, source) {
   }
 }
 
-function indentLines(content) {
-  return content
-    .split('\n')
-    .map(it => `  ${it}`)
-    .join('\n');
-}
-
 function ensureImport(root, name, source) {
   let body = root.get().value.program.body;
 
@@ -225,7 +218,6 @@ function removeImport(root, name, source) {
       .remove();
 
     if (declaration.get().value.specifiers.length === 0) {
-      // console.log(root.get('program', 'body'))
       declaration.remove();
     }
   }
@@ -278,5 +270,4 @@ module.exports = {
   removeDecorator,
   ensureImport,
   removeImport,
-  indentLines,
 };
