+++ lib/utils/template.js
@@ -1,0 +1,10 @@
+function indentLines(content) {
+  return content
+    .split('\n')
+    .map(it => `  ${it}`)
+    .join('\n');
+}
+
+module.exports = {
+  indentLines,
+};
